fix env
